// src/types/product.ts
export interface Product {
  id: number
  title: string
  price: number
  description: string
  category: string
  image: string
  rating: {
    rate: number
    count: number
  }
}

export type SortOption = 'none' | 'price-asc' | 'price-desc'

export interface FilterState {
  searchQuery: string
  selectedCategory: string
  sortBy: SortOption
}